// src/app/models/UserPreferences.ts
/**
 * UserPreferences.ts
 *
 * Created by jake
 * Created on 2/27/18
 *
 * Model user preferences.
 */


export default class UserPreferences {
  uuid: string;
  abv_low: string;
  abv_hi: string;
  ibu_low: string;
  ibu_hi: string;
  like_description: string;
  user: string;

  constructor() {

  }
}